refactor: store type 파일내에서 정의로 수정
