+++ src/app/_stores/client/modalStore.ts
@@ -21,4 +21,15 @@ const useModalStore = create(
   ),
 );
 
+interface ModalState {
+  isOpen: boolean;
+  modalName: string;
+  selectedImage: string;
+  openModal: () => void;
+  closeModal: () => void;
+  setModal: (state: string) => void;
+  setSelectedImage: (state: string) => void;
+  resetModalState: () => void;
+}
+
 export default useModalStore;
