--- src/app/_types/feed_type.ts
@@ -1,7 +1,0 @@
-import { FeedProps } from "@/_components/molecules/feed/Feed";
-
-export interface FeedState {
-  feedList: FeedProps[];
-  setFeedsState: (state: FeedProps[]) => void;
-  resetFeedState: () => void;
-}
