--- src/app/_types/modal_type.ts
@@ -1,10 +1,0 @@
-interface ModalState {
-  isOpen: boolean;
-  modalName: string;
-  selectedImage: string;
-  openModal: () => void;
-  closeModal: () => void;
-  setModal: (state: string) => void;
-  setSelectedImage: (state: string) => void;
-  resetModalState: () => void;
-}
